pyproject.toml
[project]
name = "general-life-tasks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "duckdb==1.3.2",
    "limits>=4.2",
    "loguru==0.7.3",
    "mistralai==1.9.10",
    "polars==1.33.0",
    "prometheus-client>=0.22.1",
    "psycopg2-binary>=2.9.11",
    "pydantic==2.11.7",
    "pydantic-core>=2.33.2",
    "pydantic-settings==2.10.1",
    "python-dotenv==1.1.1",
    "python-youtube>=0.9.8",
    "pytz==2025.2",
    "scrapetube>=2.6.0",
    "sqlalchemy>=2.0.44",
    "ynab>=1.8.0",
    "youtube-transcript-api>=1.2.3",
]

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
